pyproject.toml
[build-system]
requires = ["setuptools>=69", "wheel"]
build-backend = "setuptools.build_meta"

[project]
name = "landmark-demo"
version = "0.1.0"
description = "Sprint 1 Jongno landmark recognition demo (Streamlit + MobileCLIP2-S4)"
requires-python = ">=3.10"
dependencies = [
    "streamlit>=1.30",
    "pillow>=10.0",
    "numpy>=1.26",
    "tomli>=2.0; python_version<'3.11'",
    "torch>=2.3",
    "open-clip-torch>=2.30",
    "huggingface_hub>=0.24",
]

[project.optional-dependencies]
onnx = [
    "onnx>=1.16",
    "onnxruntime>=1.18",
]
dev = [
    "pytest>=7.0",
]

[tool.setuptools.packages.find]
where = ["src"]
